requirements.txt
Flask
python-docx